split up trainer
